[BF] SUBLEARN-52 edit project logic

+++ src/main/java/ru/zagamaza/competition/infra/dao/repository/BaseRepository.java
@@ -4,12 +4,12 @@ import ru.zagamaza.competition.infra.dao.entity.Entity;
 
 public interface BaseRepository<E extends Entity> {
 
-    E get(Integer id);
+    E get(Long id);
 
     E create(E entity);
 
     E update(E entity);
 
-    void delete(Integer id);
+    void delete(Long id);
 
 }
