+++ src/main/java/ru/zagamaza/competition/infra/service/BaseResourceInfraService.java
@@ -3,13 +3,14 @@ package ru.zagamaza.competition.infra.service;
 import ru.zagamaza.competition.domain.model.DomainModel;
 
 public interface BaseResourceInfraService<M extends DomainModel> {
-    M get(Integer id);
+
+    M get(Long id);
 
     M create(M model);
 
     M update(M resource);
 
-    void delete(Integer id);
+    void delete(Long id);
 
 
 }
